+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fixed an issue where **validate** failed to recognize integration tests that were missing from config.json
 * Added support for **xpanse** marketplace in **create-id-set** and **create-content-artifacts** commands.
 * Fixed an issue where **split** failed on yml files.
+* Added support for marketplace-specific tags.
 
 ## 1.7.7
 * Fixed an issue where paybooks **generate-docs** didn't parse complex input values when no accessor field is given correctly.